apps/lesson-plan-generator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
openai>=1.30.0
python-dotenv>=1.0.0
mutagen>=1.47.0
pydantic>=2.0.0
